backend/requirements.txt
# Backend Requirements

# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
aiosqlite
greenlet

# Supabase (Alternative to direct PostgreSQL)
supabase==2.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Redis for caching and rate limiting
redis==5.0.1

# HTTP client for external APIs
httpx>=0.24,<0.26

# Razorpay
razorpay==1.4.1

# SMS (Twilio)
twilio==9.0.0

# Email
aiosmtplib>=2.0

# Monitoring
sentry-sdk[fastapi]>=1.40

# Utilities
python-dotenv==1.0.0
phonenumbers==8.13.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
